# user-management/src/main/resources/openapi/collaborator.yaml
openapi: 3.0.0
info:
  title:   Collaborator Services API
  version: '1.0.0'
  description: API for managing Collaborators in the Makani system
paths:
  '/collaborators':
    get:
      summary: Retrieve a list of Collaborators
      operationId: getCollaborators
      responses:
        '200':
          description: A list of Collaborators
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GetCollaboratorResponse'
        '500':
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: './commons.yaml#/components/schemas/ErrorResponse'
    post:
      summary: Create a new Collaborator
      operationId: createCollaborator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollaboratorCreationRequest'
      responses:
        '201':
          description: Collaborator created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollaboratorCreationResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: './commons.yaml#/components/schemas/ErrorResponse'
        '500':
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: './commons.yaml#/components/schemas/ErrorResponse'
  '/collaborators/{collaboratorId}':
    get:
      summary: Retrieve an Collaborator by ID
      operationId: getCollaboratorById
      parameters:
        - in: path
          name: collaboratorId
          required: true
          schema:
            type: integer
          description: ID of the Collaborator to retrieve
      responses:
        '200':
          description: Collaborator retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCollaboratorResponse'
        '404':
          description: Collaborator not found
          content:
            application/json:
              schema:
                $ref: './commons.yaml#/components/schemas/ErrorResponse'
        '500':
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: './commons.yaml#/components/schemas/ErrorResponse'
    delete:
      summary: Delete an collaborator by ID
      operationId: deleteCollaboratorById
      parameters:
        - in: path
          name: collaboratorId
          schema:
            type: integer
          required: true
          description: ID of the collaborator to delete
      responses:
        '204':
          description: collaborator deleted successfully
        '404':
          description: collaborator not found
          content:
            application/json:
              schema:
                $ref: './commons.yaml#/components/schemas/ErrorResponse'
        '500':
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: './commons.yaml#/components/schemas/ErrorResponse'
components:
  schemas:
    BaseCollaborator:
      allOf:
        - $ref: './commons.yaml#/components/schemas/BasePerson'
        - type: object
          properties:
            skills:
                type: array
                items:
                  type: string
                  maxLength: 100
                description: list of skills
            profilePicture:
              type: string
              format: byte
      required:
        - skills
        - profilePicture
    CollaboratorCreationRequest:
      allOf:
        - $ref: '#/components/schemas/BaseCollaborator'
        - type: object
          properties:
            internalAuth:
              $ref: './commons.yaml#/components/schemas/InternalAuth'
    CollaboratorCreationResponse:
      type: object
      properties:
        collaboratorId:
          type: integer
          description: Collaborator ID
          readOnly: true
      required:
        - collaboratorId
    GetCollaboratorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseCollaborator'
        - type: object
          properties:
            collaboratorId:
              type: integer
              description: Collaborator ID
      required:
        - collaboratorId